Further sparsify and change hash function
- Sparsify past the first iteration
- Prepare for multithreaded sparsification using Givens rotations on
  independent pairs of lines
- Change hash function to avoid collisions (which appeared now)
- Added fine-tuning generated random circuit in Python test, with
  proportions of CNOTs and unitaries. Those makes the circuit more
realistic and gives the advantage to sparse simulation like in
Superpositeur.

Results:
I finally get much faster results than Qiskit's density matrix
computation. This is with CNOTs and unitaries instead of only random
Kraus.

I need to run the profiler because when there are too many random Kraus
operators, runtime explodes. This is probably due to the sparsification
algo which is likely slow. This will be parallelized soon.